fix(desktop): let named bundles reuse seeded auth
Keep forced startup refresh for production and Beta, while allowing named non-production bundles to use an unexpired owner-checked token seeded from a signed app. AuthService still refreshes normally on expiry.

Verified: xcrun swift test --package-path Desktop --filter AgentRuntimeProcessTests; desktop/macos/scripts/agent-logic-harness.sh --cross-surface-smoke.

Failure-Class: none
