+++ desktop/macos/Desktop/Sources/Chat/AgentRuntimeProcess.swift
@@ -82,36 +82,6 @@ enum AgentRuntimeStartupAdmission {
   }
 }
 
-/// Firebase credentials are mandatory for every production and model runtime
-/// start. The sole exception is a non-production journal-control start, whose
-/// owner-bound RPCs deliberately operate without a model credential. The
-/// fault suite supplies a separate, inert model token so its named test bundle
-/// can reach the local 5xx endpoint without contacting Firebase.
-enum AgentRuntimeCredentialPolicy {
-  static let hermeticFaultModelTokenEnvironmentKey = "OMI_FAULT_MODEL_AUTH_TOKEN"
-  static let hermeticFaultBundleIdentifier = "com.omi.omi-fault"
-
-  static func hermeticFaultModelToken(
-    isNonProduction: Bool,
-    bundleIdentifier: String,
-    environment: [String: String] = ProcessInfo.processInfo.environment
-  ) -> String? {
-    guard isNonProduction, bundleIdentifier == hermeticFaultBundleIdentifier else { return nil }
-    let token =
-      environment[hermeticFaultModelTokenEnvironmentKey]?
-      .trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-    return token.isEmpty ? nil : token
-  }
-
-  static func requiresManagedCredentials(
-    requestedCredentials: Bool,
-    isNonProduction: Bool,
-    hermeticFaultModelToken: String? = nil
-  ) -> Bool {
-    (requestedCredentials || !isNonProduction) && hermeticFaultModelToken == nil
-  }
-}
-
 /// Serializes the pipe read and sequence assignment performed by Foundation's
 /// readability callback. The callback may be re-entered on different threads;
 /// sequencing only after `availableData` would still allow a later read to be
@@ -2629,7 +2599,11 @@ actor AgentRuntimeProcess {
         isNonProduction: AppBuild.isNonProduction,
         hermeticFaultModelToken: hermeticFaultModelToken)
     let authService = await MainActor.run { AuthService.shared }
-    let forceRefreshToken = preferredAdapterId == .piMono && !DesktopLocalProfile.isEnabled
+    let forceRefreshToken =
+      preferredAdapterId == .piMono
+      && AgentRuntimeCredentialPolicy.shouldForceRefreshAtStartup(
+        isNonProduction: AppBuild.isNonProduction,
+        isDesktopLocalProfile: DesktopLocalProfile.isEnabled)
     let authHeader = try? await Self.startupAuthHeader(
       requiresCredentials: requiresPiMonoCredentials,
       fetchAuthHeader: {
